Make checksum an unsigned short

+++ net/udp/udp.h
@@ -9,7 +9,7 @@
 typedef struct {
     uint8_t src,
     uint8_t dst,
-    uint8_t checksum,
+    unsigned short checksum,
     unsigned short length,
     uint8_t *data_octets
 } UDP_HEADER_T;
@@ -25,7 +25,7 @@ public:
     virtual RetType transmit(Packet &packet, sockmsg_t &info, NetworkLayer *caller) = 0;
 
 private:
-    uint32_t calc_checksum(uint8_t *payload) {
+    unsigned short calc_checksum(uint8_t *payload) {
         int checksum = 0;
         for (int i = 0; i < sizeof(payload); i++) {
             checksum += payload[i];
@@ -34,7 +34,7 @@ private:
         return checksum;
     }
 
-    bool verify_checksum(uint8_t *payload, int checksum) {
+    bool verify_checksum(uint8_t *payload, unsigned short checksum) {
         int result;
 
         for (int i = 0; i < sizeof(payload); i++) {
